+++ code/sys/http/deno.json
@@ -6,14 +6,12 @@
     "lint": "deno lint",
     "dry": "deno publish --allow-dirty --dry-run",
     "clean": "deno run -RWE ./scripts/-clean.ts",
-    "test": "deno test -RWNE --allow-run --allow-ffi",
-    "dev": "deno run -RWNE --allow-run --allow-ffi ./scripts/-dev.ts",
-    "build": "deno run -RWNE --allow-run --allow-ffi ./scripts/-build.ts",
-    "serve": "cd dist && serve"
+    "test": "deno test -RWNE --allow-run --allow-ffi"
   },
   "exports": {
     ".": "./src/mod.ts",
     "./t": "./src/types.ts",
-    "./types": "./src/types.ts"
+    "./types": "./src/types.ts",
+    "./client": "./src/m.Client/mod.ts"
   }
 }
