+++ code/sys/http/src/-test/mod.ts
@@ -1,2 +1,2 @@
-export { describe, expect, expectError, it, Testing } from '@sys/std/testing';
+export { describe, expect, expectError, it, Testing } from '@sys/std/testing/server';
 export * from '../common.ts';
