+++ code/sys/http/src/common/t.ts
@@ -1,2 +1,4 @@
+export type { HttpUrl, UrlLib, StdIsLib } from '@sys/std/t';
 export type * from '@sys/types';
+
 export type * from '../types.ts';
