+++ code/sys/http/src/m.Client/mod.ts
@@ -1,0 +1,6 @@
+/**
+ * @module
+ * HTTP client tools.
+ */
+export { Fetch } from '../m.Fetch/mod.ts';
+export { HttpClient } from './m.HttpClient.ts';
