+++ code/sys/http/src/m.Fetch/common.ts
@@ -1,0 +1,1 @@
+export * from '../common.ts';
