+++ code/sys/http/src/m.Fetch/mod.ts
@@ -1,0 +1,6 @@
+/**
+ * @module
+ * Helpers for working with the native `fetch` function.
+ * https://fetch.spec.whatwg.org/
+ */
+export { Fetch } from './m.Fetch.ts';
