+++ code/sys/http/src/m.Http/common.ts
@@ -1,0 +1,2 @@
+export * from '../common.ts';
+export { Fetch } from '../m.Fetch/mod.ts';
