+++ code/sys/http/src/m.Http/m.Http.ts
@@ -1,9 +1,7 @@
-import type { t } from './common.ts';
+import { type t, Url } from './common.ts';
 
-import { HttpClient as Client } from '../m.Http.Client/mod.ts';
-import { Fetch } from '../m.Http.Fetch/mod.ts';
-import { Is } from './m.Is.ts';
-import { HttpUrl as Url } from './m.Url.ts';
+import { HttpClient as Client } from '../m.Client/mod.ts';
+import { Fetch } from '../m.Fetch/mod.ts';
 import { toError, toHeaders, toResponse } from './u.ts';
 
 /**
@@ -12,7 +10,6 @@ import { toError, toHeaders, toResponse } from './u.ts';
 export const Http: t.HttpLib = {
   Fetch,
   Client,
-  Is,
   Url,
 
   client: Client.create,
