+++ code/sys/http/src/m.Http/mod.ts
@@ -5,5 +5,5 @@
  *  - Standards body: Internet Engineering Task Force (IETF)
  *  - HTTP Working Group (HTTP-WG).
  */
-export { Fetch } from '../m.Http.Fetch/mod.ts';
+export { Fetch } from '../m.Fetch/mod.ts';
 export { Http } from './m.Http.ts';
