+++ code/sys/http/src/mod.ts
@@ -4,8 +4,8 @@
  *
  * @example
  * ```ts
- * import { pkg } from '@sys/http';
- * import { Http } from '@sys/http/client';
+ * import { pkg } from 'jsr:@sys/http';
+ * import { Http } from 'jsr:@sys/http/client';
  *
  * const a = Http.Client.create({ accessToken: 'my-jwt' });
  * const b = Http.client();  // ← shorthand.
@@ -15,3 +15,8 @@ export { pkg } from './pkg.ts';
 
 /** Module types. */
 export type * as t from './types.ts';
+
+/**
+ * Library
+ */
+export { Http, Fetch } from './m.Http/mod.ts';
