--- code/sys/std/src/m.Http.Client/mod.ts
@@ -1,1 +1,0 @@
-export { HttpClient } from './m.HttpClient.ts';
