--- code/sys/std/src/m.Http.Fetch/common.ts
@@ -1,3 +1,0 @@
-export * from '../common.ts';
-export { Err } from '../m.Err/mod.ts';
-export { rx } from '../m.Rx/mod.ts';
