--- code/sys/std/src/m.Http.Fetch/mod.ts
@@ -1,1 +1,0 @@
-export { Fetch } from './m.Fetch.ts';
