--- code/sys/std/src/m.Http/common.ts
@@ -1,2 +1,0 @@
-export * from '../common.ts';
-export { Fetch } from '../m.Http.Fetch/mod.ts';
